server/requirements/base.txt
Django==1.11.2
django-rest-auth==0.9.1
django-allauth==0.32.0
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
psycopg2==2.7.1
PyJWT==1.5.2
